Add rake task for scaling up and down
These could be run as regularly scheduled jobs. The jobs
should be scaling the workers up and down themselves, but
this would pick up any scaling that might have been missed.
(e.g., if the Heroku API was down or if the worker restarted
while trying to scale, or whatever.)